refactor: extract requireRole helper from devtoken
